Rewriting code part 1
