--- comum.py
@@ -1,4 +1,0 @@
-''' Variavéis comuns '''
-PATH_TEMPLATE_HAS = '/data/templates'
-PATH_TEMPLATE = 'templates'
-FILE_COMM = 'comm.json'
